Add an implementation of murmurhash for appengine
Appengine cant use unsafe so it cant use the murmur implementation
add an implementation which does not use unsafe protected by the
appengine build flag.

fixes #457
